Create separate display component for quiz score
* increase readability of the code
* improve maintainability
