+++ src/components/QuizApp.tsx
@@ -1,28 +1,11 @@
 import React, { useState } from "react";
 import "../styles/QuizApp.css";
 import { type Question, retrieveQuizQuestions } from "../data/questions.ts";
+import { QuizScoreDisplay } from "./QuizScoreDisplay.tsx";
 
 // non-negative values correspond to question index
 const QUIZ_OPENING_STATE = -1;
 const QUIZ_CLOSING_STATE = -2;
-
-const closingStatement = (score: number): string => {
-  if (score <= 2)
-    return "Don’t worry, Rookie Lobstar! Every Lobstar starts somewhere, and you’re just getting your feet wet. Keep learning, practicing, and throwing! You’ll be lighting up the field in no time.";
-
-  if (score <= 4)
-    return "Getting there, Aspiring Lobstar! You’ve made some good catches here. Dive back into the playbook and sharpen your game—you’re on the way to greatness!";
-
-  if (score <= 6)
-    return "Nice one, Steady Lobstar! You’re showing solid skills and knowledge out there. A few more reps and you’ll be skying for the disc like a pro.";
-
-  if (score <= 8)
-    return "Impressive, Rising Lobstar! You’ve clearly got a good grip on the game. Keep hustling and aiming for the stars—your ultimate future looks bright.";
-
-  return "Outstanding, All-Star Lobstar! You’ve nailed it like a perfect hammer throw. Your knowledge is championship material—just like our Lobstars spirit on the field.";
-};
-
-
 
 const QuizApp: React.FC = () => {
   const [currentQuestionIndex, setCurrentQuestionIndex] =
@@ -62,7 +45,12 @@ const QuizApp: React.FC = () => {
   if (currentQuestionIndex === QUIZ_OPENING_STATE) {
     return (
       <div className="quiz-container">
-        <p>Welcome, Lobstar! This quiz is designed to enhance your knowledge of Ultimate Frisbee and help you grow as a valued member of the Lobstars team. Test your skills, learn something new, and aim for the stars. Good luck!</p>
+        <p>
+          Welcome, Lobstar! This quiz is designed to enhance your knowledge of
+          Ultimate Frisbee and help you grow as a valued member of the Lobstars
+          team. Test your skills, learn something new, and aim for the stars.
+          Good luck!
+        </p>
         <button onClick={startQuiz}>Start Quiz</button>
       </div>
     );
@@ -72,10 +60,9 @@ const QuizApp: React.FC = () => {
     return (
       <div className="quiz-container">
         <h2 className="text-4xl">Quiz Completed!</h2>
-        <p id="score">
-          Your score: {score} / {questions.length}
-        </p>
-        <p>{closingStatement(score)}</p>
+
+        <QuizScoreDisplay score={score} totalScore={questions.length} />
+
         <button onClick={startQuiz}>Restart Quiz</button>
       </div>
     );
@@ -84,70 +71,77 @@ const QuizApp: React.FC = () => {
   const currentQuestion = questions[currentQuestionIndex];
 
   return (
-      <div className="quiz-container">
-
-        <div className="pb-4" id="container-for-progress-bar">
-          <div className="absolute top-1 right-3" id="progress-bar">
-            <label className="text-sm" htmlFor="quiz-progress">Quiz progress: </label>
-            <progress id="quiz-progress" max={questions.length} value={currentQuestionIndex}> </progress>
-          </div>
+    <div className="quiz-container">
+      <div className="pb-4" id="container-for-progress-bar">
+        <div className="absolute right-3 top-1" id="progress-bar">
+          <label className="text-sm" htmlFor="quiz-progress">
+            Quiz progress:{" "}
+          </label>
+          <progress
+            id="quiz-progress"
+            max={questions.length}
+            value={currentQuestionIndex}
+          >
+            {" "}
+          </progress>
         </div>
+      </div>
 
-        <h2>{currentQuestion.question}</h2>
-        <div>
-          {currentQuestion.answers.map((answer: string, index: number) => (
-              <button
-                  key={index}
-                  onClick={() => setSelectedAnswer(index)}
-                  className={`answer-btn \
+      <h2>{currentQuestion.question}</h2>
+      <div>
+        {currentQuestion.answers.map((answer: string, index: number) => (
+          <button
+            key={index}
+            onClick={() => setSelectedAnswer(index)}
+            className={`answer-btn \
 ${isSubmitted && index === currentQuestion.correctIndex ? "correct" : ""} \
 ${isSubmitted && index === selectedAnswer && index !== currentQuestion.correctIndex ? "incorrect" : ""} \
 ${!isSubmitted && index === selectedAnswer ? "selected" : ""}\
 `}
-                  disabled={isSubmitted}
-              >
-                {answer}
-              </button>
-          ))}
-        </div>
-        {isSubmitted ? (
-            <div>
-              <p>
-                {selectedAnswer === currentQuestion.correctIndex
-                    ? "Correct!"
-                    : "Incorrect!"}
-              </p>
-              <p>{currentQuestion.explanation}</p>
-              <p>
-                <small>
-                  Reference:{" "}
-                  <a
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      href={currentQuestion.reference.link}
-                  >
-                    {" "}
-                    {currentQuestion.reference.sourceName}
-                  </a>{" "}
-                  ({currentQuestion.reference.locator})
-                </small>
-              </p>
-              <button className={"question-change-btn"} onClick={nextQuestion}>
-                {currentQuestionIndex < questions.length - 1
-                    ? "Next Question"
-                    : "Finish Quiz"}
-              </button>
-            </div>
-        ) : (
-            <button
-                className={"submit-btn"}
-                onClick={submitAnswer}
-                disabled={selectedAnswer === null}
-            >
-              Submit
-            </button>
-        )}
+            disabled={isSubmitted}
+          >
+            {answer}
+          </button>
+        ))}
       </div>
+      {isSubmitted ? (
+        <div>
+          <p>
+            {selectedAnswer === currentQuestion.correctIndex
+              ? "Correct!"
+              : "Incorrect!"}
+          </p>
+          <p>{currentQuestion.explanation}</p>
+          <p>
+            <small>
+              Reference:{" "}
+              <a
+                target="_blank"
+                rel="noopener noreferrer"
+                href={currentQuestion.reference.link}
+              >
+                {" "}
+                {currentQuestion.reference.sourceName}
+              </a>{" "}
+              ({currentQuestion.reference.locator})
+            </small>
+          </p>
+          <button className={"question-change-btn"} onClick={nextQuestion}>
+            {currentQuestionIndex < questions.length - 1
+              ? "Next Question"
+              : "Finish Quiz"}
+          </button>
+        </div>
+      ) : (
+        <button
+          className={"submit-btn"}
+          onClick={submitAnswer}
+          disabled={selectedAnswer === null}
+        >
+          Submit
+        </button>
+      )}
+    </div>
   );
 };
